feedback_site/requirements.txt
Flask==2.2.3
Pillow==9.3.0
psycopg2==2.9.5
validate_email==1.3
Werkzeug==2.2.3
